requirements.txt
feedparser==6.0.10
requests==2.31.0
rich==13.7.0
groq
newspaper3k
lxml[html_clean]